restyled CallToAction.tsx
