+++ components/action/CallToActionButton.tsx
@@ -4,7 +4,7 @@ import CustomButton from '../reusable/CustomButton';
 
 const CallToActionButton = () => {
   return (
-    <Link className="md:mr-[2.25rem]" href="/work">
+    <Link className="" href="/work">
       <CustomButton title="Get in touch with me" />
     </Link>
   );
