+++ components/action/CallToActionHeader.tsx
@@ -1,6 +1,6 @@
 const CallToActionHeader = () => {
   return (
-    <h3 className="mb-[1rem] flex max-w-[19.6rem] flex-wrap justify-center text-[1.875rem] font-bold text-black-200 sm:w-1/2">
+    <h3 className="flex pb-4 text-3xl font-bold  leading-[105%] tracking-[-0.009rem] text-black-200 sm:max-w-[32rem] plusLarge:pb-[5.125rem]  plusLarge:text-5xl  plusLarge:leading-[115%] plusLarge:tracking-[0.03rem]">
       Have a project in mind that requires technical expertise?
     </h3>
   );
