+++ components/reusable/AlexanderBrand.tsx
@@ -1,0 +1,221 @@
+import { IconType } from '@/types';
+
+export const AlexanderBrand = ({
+  className,
+  height = 80,
+  width = 80,
+}: IconType) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    version="1.1"
+    className={className}
+    width={width}
+    height={height}
+    viewBox="0 0 1000 1000"
+  >
+    <rect
+      className="fill-white-800 dark:fill-black-300"
+      width="80"
+      height="80"
+    ></rect>
+    <g transform="matrix(0.7,0,0,0.7,150.00000449757903,106.25)">
+      <svg
+        className="fill-black-300 dark:fill-white-900"
+        viewBox="0 0 320 360"
+        preserveAspectRatio="xMidYMid meet"
+        height="1125"
+        width="1000"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <g
+          id="tight-bounds"
+          transform="matrix(1,0,0,1,0.0000020560361235766322,0)"
+        >
+          <svg viewBox="0 0 320 360" height="360" width="320">
+            <g>
+              <svg></svg>
+            </g>
+            <g>
+              <svg viewBox="0 0 320 360" height="360" width="320">
+                <g>
+                  <path
+                    className="fill-primary-light dark:fill-primary-dark"
+                    transform="translate(-15.58119,4.24857) scale(3.51065,3.51065)"
+                    d="M51.5 85.2c-0.4-0.5-0.8-0.2-0.8-0.5-2.4 0.6-6.9-1-8.8-0.8 0.4-0.6-0.6-0.5-0.6-0.9-0.4 0.6-0.8 0-1.4 0.2-0.6-1.5-3-0.4-3.3-2.3-1.4 0-1.9-0.8-3.3-1.6 0.1-0.1 0.2-0.1 0.3-0.2-0.7-0.1-0.7-0.5-1.3-0.5l0.2-0.4c-1.1-0.2-1.5-1.9-3.1-2.5-0.5-0.1-1.6-1.8-2.2-2.2-0.3-0.2-0.8-0.2-0.1-0.3-0.1-0.1-0.1-0.3 0-0.4-0.7-0.2-0.8-0.8-1.7-1.1-0.1-0.3 0.2-0.5 0-0.7-0.2-0.1-0.5 0.1-0.7 0-0.1 1.8 1.9 2 2.5 3.5-0.2-0.2-0.5-0.8-0.7-0.4-0.1 0.3 0.8 1.4 0.8 0.8-0.3 0.2-0.4-0.3-0.2-0.4 0.3 0.5 0.9 0.4 0.9 0.8-0.3 0.3-0.7-0.3-0.6 0.2 0.9 0.5 0.8 1 1.3 1.7-0.4-0.5-0.5 0.4-0.9 0-0.1-1.5-1.8-1.2-1.8-2.7-0.2-0.5-0.4-0.1-0.7-0.3 0.4-0.4-1.2-1.8-1.2-1.3-0.5-0.6-0.2-0.8-0.3-1.2-0.4-0.1-0.1 0.4-0.4 0.1 0.4-0.7-0.5-1.6-1.1-2.3l0.4-0.2c-0.7-0.4-0.8-1.2-1.3-2.1-0.6 0.9 1.2 2.9 0.9 3.3 0 0.1 0.7 0.9 0.9 1.3 0.8 1.4 1.4 3.1 2.5 3.9-0.2-0.1-0.1-0.7 0.2-0.4 0 0.3-0.2 0.6-0.2 0.9 0.6 0.6 1.1 1.1 1.7 1.7-1.3-0.8-2.6-1.6-3.5-2.8 0.5-0.1 1 0.8 1.4 0.7-1-0.8-0.7-1.1-1.1-2.1-0.4-0.3 0.1 0.4-0.3 0.5-1.3-1.2-2.3-1.9-2.9-3 0.1-0.1 0.2-0.2 0.2-0.4-0.1-0.1-0.1-0.1-0.2-0.2 0.6 0.6 0.8 1.1 1.5 1.2-0.3-0.3 0-0.4-0.4-0.7-0.2-0.4-0.5 0.2-0.7-0.1 0.2-0.3 0-0.7 0-1-0.2 0-0.4 0-0.5-0.3-0.2-0.1-0.3 0-0.4 0.2-0.1-0.1-0.1-0.3-0.2-0.4v-0.1c-0.1-0.1-0.2-0.3-0.3-0.4-0.8-2.1-1.3-4.9-2.4-6.7 0.3-1.2-0.2-3.6-0.5-4.2 0.3-0.5 0.1-2.8-0.1-3.2-0.5 0.2 0.1 1.3-0.4 1.6-0.5-2.2 0.3-4.7 0.1-6.7 0.2 0 0.4 0 0.4-0.1-0.3-0.3 0.2-0.9 0.1-1.6 0.4-0.9-0.9 1.3-0.9 2.2-0.1 0.2-0.5 0-0.4 0.5 0 0.7 0.2 2.1-0.2 3.1-0.2 0-0.2-0.2-0.4-0.1-0.1 0.5-0.2 1.1-0.3 1.6-0.1-0.2-0.2-0.5-0.2-0.6-0.5-2.5 0-5.4-0.2-7.5 0.4 0 0.4-0.6 0.7-0.8-0.1 0.8-0.3 1.8 0.2 2 0.4-3.9 1.2-8.2 3-12.4 1.3-3.1 3.2-6.2 5.5-8.7 0.2 0.3-0.1 1 0 1.3 2-2.1 3.3-4.6 5.5-6 0.1 0.5-0.9 0.9-0.9 1.3 0.9-0.9 1.2-0.6 2.2-0.9 0.3-0.4-0.4 0-0.5-0.4 0.6-0.4 1-0.9 1.4-1.2 0 0.1-0.1 0.1-0.1 0.2 1.1-0.3 1.2-0.9 2-1.1-0.1-0.1-0.3-0.2-0.6-0.1 0.2-0.1 0.5-0.3 0.8-0.4-0.1 0-0.1 0.1-0.2 0.1 0.6-0.3 0.1 0.1 0.1 0.3 0.1 0 0.2-0.1 0.2-0.1-0.6 0.4-1 0.6-1.2 1.3 0.3-0.2 0.4 0 0.8-0.3 0.4-0.2-0.1-0.5 0.2-0.7 0.3 0.2 0.7 0.1 1 0.1 0-0.2 0.1-0.4 0.3-0.5 0.1-0.2 0-0.3-0.1-0.3 0.1-0.1 0.2-0.1 0.4-0.2 0 0 0.1 0 0.1 0.1 1.7-0.8 3.6-1.5 5.4-2.1 1.1-0.1 2.5-0.1 3.1-0.8-0.2 0-0.4 0.1-0.5 0 0.1 0 0.3-0.1 0.4-0.1 0.2 0.2 0.5 0 0.2 0.4 0.1 0 0.1 0 0.2-0.1 0 0.1 0 0.1 0.1 0.2 0.1 0 0.2-0.1 0.3-0.3 0.1 0 0.2 0 0.3-0.1l-0.1 0.1c0.8-0.2 1.5-0.1 2.6-0.3 0-0.1 0-0.2 0-0.3 0.2 0 0.3-0.1 0.5-0.1v0.1c0.4-0.2 0.8 0.2 0.8 0 0.1 0 0.1 0 0.2 0 0 0.2-0.1 0.3-0.1 0.4 0.5 0 0.8 0 1-0.3-0.1-0.1-0.2-0.1-0.3-0.1 0.2 0 0.4-0.1 0.6-0.2 0.3 0.1 0.4 0.2 0.4 0.6 0.5-0.9 2.1 0.2 2.8 0 0.3 0 0.2-0.4 0.4-0.6 0.1 0 0.2 0 0.3 0 0.5 0.4 1.2 0.8 2.1 0.9-0.8-0.2 0.1-0.3 0.2-0.4-0.2 0-0.2-0.1-0.2-0.2 0.3 0 0.6 0.1 0.9 0.1 0 0.2 0.1 0.3 0.3 0.3 0.2 0 0.2-0.1 0.2-0.3 0.3 0 0.5 0 0.8 0.1-0.1 0.4 0.5 0.5 0.7 0.8-0.8-0.2-1.7-0.5-2 0 3.2 0.7 6.7 1.7 10.1 3.4 0 0 0 0 0 0.1h0.1c0.6 0.3 1.2 0.6 1.8 0.9 0.4 0.2 0.7 0.4 1.1 0.7 0 0.1 0 0.2 0.2 0.2h0.1c0.4 0.3 0.8 0.5 1.2 0.8 0.4 0.6 0.8 1.4 1.5 1.2 1.5 2.3 3.1 3.2 4.4 5.6 0.1-0.4 0.4 0.9 0.9 1-0.8-0.1 0.5 0.5-0.3 0.4 0.7 0.7 0.9 1 1 1.8-0.2-0.4-0.3-0.5-0.4-0.1-0.1 0.6 0.7-0.1 1 0.5-0.6 0-0.4 0.4-0.3 0.9 0.3-0.3 0.2-0.6 0.5-0.5 0.4 0.4-0.2 0.5-0.1 0.8 0.3-0.4 0.4-0.2 0.8 0-0.5 0.1 0.3 0.8-0.3 0.5 0.4 0.5 0.5 0.8 0.3 1.3-0.4-0.6-0.1-1-0.7-0.9 1.2 1 0.6 3.3 2 3.5-0.9 0.2 0 0.6-0.5 0.7 0.6 0.2 0.3 1 1 1.2-0.1 0.5-0.3 0.8-0.4 1.3 0.3 0.4 0.4 1.2 0.9 1.3-0.3 0.3-0.1 1.1-0.4 1.4 0.4-0.2 0.6-0.2 0.9 0.2-0.5 0.4-0.1 0.8-0.4 0.9 0.9 2.3-0.2 7 0.3 8.8-0.6-0.3-0.4 0.7-0.8 0.7 0.7 0.3 0.1 0.8 0.4 1.3-1.4 0.8 0 3-1.9 3.5 0.2 1.4-0.6 2-1.2 3.5-0.1-0.1-0.1-0.2-0.2-0.3-0.1 0.7-0.5 0.8-0.3 1.4l-0.4-0.2c0 1.1-1.7 1.7-2.1 3.4-0.1 0.5-1.6 1.8-2 2.5-0.2 0.3-0.1 0.8-0.3 0.1-0.1 0.1-0.3 0.1-0.4 0.1-0.1 0.7-0.7 0.9-0.9 1.8-0.2 0.1-0.5-0.2-0.7 0.1 0 0.2 0.2 0.5 0.1 0.7 1.8-0.1 1.8-2.2 3.2-2.9-0.1 0.3-0.7 0.6-0.3 0.8 0.3 0 1.2-1 0.7-0.9 0.3 0.2-0.3 0.4-0.4 0.3 0.5-0.4 0.3-0.9 0.7-1 0.3 0.2-0.3 0.7 0.3 0.6 0.4-0.9 0.9-1 1.5-1.5-0.4 0.4 0.5 0.5 0.1 0.9-1.5 0.3-1 2-2.5 2.2-0.4 0.3-0.1 0.4-0.2 0.7-0.5-0.3-1.6 1.4-1.2 1.3-0.5 0.5-0.8 0.3-1.2 0.4-0.1 0.4 0.4 0.1 0.2 0.4-0.7-0.3-1.6 0.7-2.2 1.4l-0.3-0.3c-0.3 0.8-1.1 0.9-2 1.5 1 0.5 2.7-1.6 3.2-1.3 0.1 0 0.9-0.8 1.2-1.1 1.3-1 2.9-1.8 3.5-2.9-0.1 0.2-0.7 0.2-0.4-0.1 0.3-0.1 0.6 0.1 0.9 0.1 1.6-2.2 3.6-4 5.1-6.6 0.1-0.1 0.2-0.3 0.3-0.5-0.2 0.6-0.5 1.1-0.7 1.7-1.6 1.1-2.2 4.3-4.1 4.8 0.1 1.1-0.9 1.4-1.4 2.7-0.3 0.4-0.5-0.3-0.8 0.2 0.3 1.4-2.5 1.3-2.4 2.6-0.2 0.1-0.5 0.2-0.7 0.4h-0.1c-0.1 0.1-0.1 0.1-0.2 0.2-0.5 0.3-0.9 0.6-1.3 0.9-1.4 0.8-3.3 1.4-4.8 2.2v-0.1c-0.4 0.2-0.9 0.1-0.8 0.5-0.3 0.2-0.6 0.4-0.9 0.6-0.1 0-0.1 0-0.2 0 0 0-0.1-0.1-0.2-0.1-0.2 0-0.2 0-0.3 0.1-1.3 0.1-3 0.7-3.5 1.1-0.5-0.3-2.8 0.2-3.2 0.4 0.2 0.3 0.8 0 1.3 0-0.3 0.1-0.6 0.2-0.7 0.4-0.1 0-0.1 0-0.2 0-0.1-0.3-0.4 0-0.5-0.1h0.1c-0.1-0.4-0.4-0.6-0.9-0.5 0 0.2 0 0.3-0.1 0.5 0.2 0 0.8-0.3 0.8 0-0.1 0-0.1 0.1 0 0.1-0.2 0.2-0.5 0.2-0.7 0.3-0.4 0-0.9 0-1.3 0-0.1-0.1-0.2-0.1-0.4-0.2-0.1 0.1-0.3 0.1-0.4 0.2-0.6 0-1.3 0.1-1.8 0.2 0-0.2 0-0.4-0.2-0.4-0.3 0.4-0.9-0.1-1.6 0.1-0.7-0.2 0.4 0.3 1.4 0.5-0.1 0.1-0.2 0.2-0.5 0.3M37.5 86.6c-1.4-1.3-4.3-1.9-4.8-2.9-0.8-0.1-2-1.6-3-1.8 0.5 0 0.1-0.3 0-0.5-0.2 0.2-0.4 0.3-0.6 0.3 0-0.8-0.8-1.3-1.5-1.8 0.5 0.3 0.9 0.5 1.5 0.6 0-0.2 0-0.4 0.2-0.4 0.7 0.7 1.5 1.4 2.4 2.1 1.2 1.1 7.4 3.9 8.5 4.4 0 0-0.2 0.4-0.1 0.4 0.5 0.3 0.9-0.3 1 0.4-1.6 0.2-2.2-0.9-3.6-0.8M83.6 63.4v0.1 0c-0.1 0 0-0.1 0-0.1M43.2 17.1c-0.1 0.1-0.2 0.1-0.2 0.1 0-0.1 0.1-0.1 0.2-0.1M41.3 17.6c0 0.1 0 0.1 0 0 0 0.1 0 0.1 0 0 0 0.1 0 0 0 0M34.7 19.7v0c0.1-0.1 0.1-0.1 0.2-0.1-0.1 0-0.2 0.1-0.2 0.1M50.7 16.4c-0.2 0-0.3 0-0.4-0.1 0.1 0 0.2 0 0.3 0h0.1c0 0 0 0 0 0.1M52.5 16.6c-0.1-0.1-0.4-0.2-0.7-0.2 0-0.1 0.1-0.2 0.1-0.2 0.1 0 0.2 0 0.3 0 0 0 0.1 0 0.1 0.1 0 0 0 0 0-0.1 0.1 0 0.1 0 0.2 0 0.1 0.2 0.2 0.3 0 0.4M55.8 16.9c-0.5 0.1-0.8-0.1-1.3-0.4 0.3 0 0.7 0.1 1 0.2 0.1 0 0.2 0.1 0.3 0.2M91.4 31.1c-0.2 0.1-0.3 0.1-0.5-0.3-1.6-3.4-4.5-8.3-8.7-12.4-4.1-4.2-9.6-7.6-14.6-9.4 0.3 0.1 0.7 0.3 0.5 0.5-4.4-1.7-8.6-2.7-12.6-3.1-10.8-1.7-22.1 0.6-30.8 6.2 0.1 0.2 0.1 0.3-0.2 0.5-3.2 2-7.7 5.5-11.3 10.1-3.6 4.6-6.4 10.4-7.6 15.6 0.1-0.3 0.2-0.8 0.4-0.6-2.6 10.3-1.9 19.1 1.7 26.7 0.2-0.2-0.5-1.2 0.1-1.1 0.4 1.1 0.9 2.3 1.5 3.4 0.7 1.6 1.5 3.7 2.6 4.2-2.8-5.5-4.7-11.8-4.9-17.4 0.4 0.8 0.3 2.5 0.9 3.2 0.2-3.8-0.7-8.9 0.6-12.5-0.3 1.9-0.5 4.5-0.3 7.1 0 1.9 0.9 7.2 1.5 7.7-0.1-0.1-0.1-0.6 0-0.6 0.3 0.6-0.4 0.9 0.1 1 0-0.2 0.1-0.2 0.2-0.2 1 2.2 2.3 5.6 3.3 7.3 1.2 2.2 2.8 4.2 4.6 6 0 0.1 0.1 0.1 0.1 0.2 0.2 0.5 0.9 1.1 1.1 1.5 0.1 0.2-0.1 0.5 0 0.7 1.5 2.8 3.7 5.6 6.4 7.8 1.7 2 4.4 4.1 7.7 5.9 3.2 1.7 6.9 3.1 10.3 3.9 2.4 0.6 5.3 0.7 7.1 1.1 2.4 0.5 6.6-0.2 8.7-0.6 12.6-1.9 25.1-10.5 31.4-22.9 6.5-12.5 6.3-27.8 0.7-39.5"
+                    fill="#428dff"
+                    fill-rule="nonzero"
+                    stroke="none"
+                    strokeWidth="1"
+                    strokeLinecap="butt"
+                    strokeLinejoin="miter"
+                    strokeMiterlimit="10"
+                    strokeDasharray=""
+                    strokeDashoffset="0"
+                    fontFamily="none"
+                    fontWeight="none"
+                    fontSize="none"
+                    textAnchor="none"
+                    data-fill-palette-color="tertiary"
+                  ></path>
+                </g>
+                <g transform="matrix(1,0,0,1,121.1016175212632,108)">
+                  <svg
+                    viewBox="0 0 77.79676495747361 144"
+                    height="144"
+                    width="77.79676495747361"
+                  >
+                    <g>
+                      <svg
+                        viewBox="0 0 77.79676495747361 144"
+                        height="144"
+                        width="77.79676495747361"
+                      >
+                        <g>
+                          <svg
+                            viewBox="0 0 119.74900990099013 221.65262829590233"
+                            height="144"
+                            width="77.79676495747361"
+                          >
+                            <g>
+                              <svg
+                                viewBox="0 0 119.74900990099013 144"
+                                height="144"
+                                width="119.74900990099013"
+                              >
+                                <g id="textblocktransform">
+                                  <svg
+                                    viewBox="0 0 119.74900990099013 144"
+                                    height="144"
+                                    width="119.74900990099013"
+                                    id="textblock"
+                                  >
+                                    <g>
+                                      <svg
+                                        viewBox="0 0 119.74900990099013 144"
+                                        height="144"
+                                        width="119.74900990099013"
+                                      >
+                                        <g transform="matrix(1,0,0,1,0,0)">
+                                          <svg
+                                            width="119.74900990099013"
+                                            viewBox="0.34375 -39.95 33.596141802400005 40.400000000000006"
+                                            height="144"
+                                            data-palette-color="#428dff"
+                                          >
+                                            <path
+                                              className="fill-primary-light dark:fill-primary-dark"
+                                              d="M28.15-1.2L17.15-34.8 6.2-1.2 6.2-1.25Q5.9-0.5 5.38-0.03 4.85 0.45 4.1 0.45L4.1 0.45Q3.7 0.45 3.33 0.38 2.95 0.3 2.55 0.2L2.55 0.2 2.6 0.2Q-0.25-0.7 0.5-3L0.5-3 11.9-37.65Q12.15-38.45 12.78-39 13.4-39.55 14.25-39.65L14.25-39.65Q15.1-39.8 15.9-39.88 16.7-39.95 17.15-39.95L17.15-39.95Q17.6-39.95 18.4-39.88 19.2-39.8 20.05-39.7L20.05-39.7Q20.75-39.6 21.43-38.95 22.1-38.3 22.45-37.65L22.45-37.65Q25.25-29.05 28.13-20.33 31-11.6 33.8-3L33.8-3Q34.5-0.7 31.7 0.2L31.7 0.2Q31.3 0.3 30.98 0.38 30.65 0.45 30.25 0.45L30.25 0.45Q29.55 0.45 28.98-0.03 28.4-0.5 28.15-1.25L28.15-1.25 28.15-1.2ZM20.9-12L20.9-12Q20.9-10.4 19.78-9.28 18.65-8.15 17.05-8.15L17.05-8.15Q15.45-8.15 14.35-9.28 13.25-10.4 13.25-12L13.25-12Q13.25-13.6 14.35-14.73 15.45-15.85 17.05-15.85L17.05-15.85Q18.65-15.85 19.78-14.73 20.9-13.6 20.9-12Z"
+                                              opacity="1"
+                                              transform="matrix(1,0,0,1,0,0)"
+                                              data-fill-palette-color="primary"
+                                              id="text-0"
+                                            ></path>
+                                          </svg>
+                                        </g>
+                                      </svg>
+                                    </g>
+                                  </svg>
+                                </g>
+                              </svg>
+                            </g>
+                            <g transform="matrix(1,0,0,1,4.37083886138614,165.013056)">
+                              <svg
+                                viewBox="0 0 111.00733217821785 56.63957229590235"
+                                height="56.63957229590235"
+                                width="111.00733217821785"
+                              >
+                                <g>
+                                  <svg
+                                    xmlns="http://www.w3.org/2000/svg"
+                                    version="1.1"
+                                    x="0"
+                                    y="0"
+                                    viewBox="5.000499999999999 27.03905580104133 89.99999999999999 45.920944198958665"
+                                    enable-background="new 0 0 100 100"
+                                    height="56.63957229590235"
+                                    width="111.00733217821785"
+                                    data-fill-palette-color="accent"
+                                    id="icon-0"
+                                  >
+                                    <g
+                                      className="fill-primary-light dark:fill-primary-dark"
+                                      data-fill-palette-color="accent"
+                                    >
+                                      <g
+                                        className="fill-primary-light dark:fill-primary-dark"
+                                        data-fill-palette-color="accent"
+                                      >
+                                        <g
+                                          className="fill-primary-light dark:fill-primary-dark"
+                                          data-fill-palette-color="accent"
+                                        >
+                                          <g
+                                            className="fill-primary-light dark:fill-primary-dark"
+                                            data-fill-palette-color="accent"
+                                          >
+                                            <path
+                                              className="fill-primary-light dark:fill-primary-dark"
+                                              d="M75.195 72.96c-0.619 0-1.241-0.21-1.75-0.64-1.146-0.967-1.291-2.681-0.324-3.827l15.609-18.493L73.122 31.507c-0.967-1.146-0.822-2.859 0.324-3.827 1.146-0.968 2.859-0.823 3.827 0.323L94.36 48.248c0.854 1.012 0.854 2.492 0 3.503L77.272 71.995C76.735 72.632 75.968 72.96 75.195 72.96z"
+                                              data-fill-palette-color="accent"
+                                            ></path>
+                                          </g>
+                                        </g>
+                                      </g>
+                                      <g
+                                        className="fill-primary-light dark:fill-primary-dark"
+                                        data-fill-palette-color="accent"
+                                      >
+                                        <g
+                                          className="fill-primary-light dark:fill-primary-dark"
+                                          data-fill-palette-color="accent"
+                                        >
+                                          <g
+                                            className="fill-primary-light dark:fill-primary-dark"
+                                            data-fill-palette-color="accent"
+                                          >
+                                            <path
+                                              className="fill-primary-light dark:fill-primary-dark"
+                                              d="M24.805 72.96c-0.772 0-1.539-0.328-2.076-0.964L5.641 51.752c-0.854-1.012-0.854-2.492 0-3.503l17.088-20.245c0.967-1.146 2.681-1.291 3.827-0.324 1.146 0.967 1.291 2.681 0.324 3.827L11.269 50.001l15.609 18.492c0.967 1.146 0.823 2.859-0.323 3.827C26.045 72.749 25.423 72.96 24.805 72.96z"
+                                              data-fill-palette-color="accent"
+                                            ></path>
+                                          </g>
+                                        </g>
+                                      </g>
+                                      <g
+                                        className="fill-primary-light dark:fill-primary-dark"
+                                        data-fill-palette-color="accent"
+                                      >
+                                        <path
+                                          className="fill-primary-light dark:fill-primary-dark"
+                                          d="M37.325 72.96c-0.492 0-0.99-0.134-1.438-0.414-1.271-0.796-1.657-2.471-0.861-3.742L60.37 28.315c0.796-1.271 2.472-1.656 3.743-0.861 1.271 0.796 1.657 2.471 0.861 3.742L39.63 71.685C39.115 72.508 38.23 72.96 37.325 72.96z"
+                                          data-fill-palette-color="accent"
+                                        ></path>
+                                      </g>
+                                    </g>
+                                  </svg>
+                                </g>
+                              </svg>
+                            </g>
+                            <g></g>
+                          </svg>
+                        </g>
+                      </svg>
+                    </g>
+                  </svg>
+                </g>
+              </svg>
+            </g>
+            <defs></defs>
+          </svg>
+          <rect
+            width="320"
+            height="360"
+            fill="none"
+            stroke="none"
+            visibility="hidden"
+          ></rect>
+        </g>
+      </svg>
+    </g>
+  </svg>
+);
