--- components/reusable/opengraph-image.tsx
@@ -1,242 +1,0 @@
-import { IconType } from '@/types';
-
-export const AlexanderBrand = ({
-  className,
-  height = 80,
-  width = 80,
-}: IconType) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    version="1.1"
-    className={className}
-    width={width}
-    height={height}
-    viewBox="0 0 1000 1000"
-  >
-    <rect
-      className="fill-white-800 dark:fill-black-300"
-      width="80"
-      height="80"
-    ></rect>
-    <g transform="matrix(0.7,0,0,0.7,150.00000449757903,106.25)">
-      <svg
-        className="fill-black-300 dark:fill-white-900"
-        viewBox="0 0 320 360"
-        preserveAspectRatio="xMidYMid meet"
-        height="1125"
-        width="1000"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <g
-          id="tight-bounds"
-          transform="matrix(1,0,0,1,0.0000020560361235766322,0)"
-        >
-          <svg viewBox="0 0 320 360" height="360" width="320">
-            <g>
-              <svg></svg>
-            </g>
-            <g>
-              <svg viewBox="0 0 320 360" height="360" width="320">
-                <g>
-                  <path
-                    className="fill-primary-light dark:fill-primary-dark"
-                    transform="translate(-15.58119,4.24857) scale(3.51065,3.51065)"
-                    d="M51.5 85.2c-0.4-0.5-0.8-0.2-0.8-0.5-2.4 0.6-6.9-1-8.8-0.8 0.4-0.6-0.6-0.5-0.6-0.9-0.4 0.6-0.8 0-1.4 0.2-0.6-1.5-3-0.4-3.3-2.3-1.4 0-1.9-0.8-3.3-1.6 0.1-0.1 0.2-0.1 0.3-0.2-0.7-0.1-0.7-0.5-1.3-0.5l0.2-0.4c-1.1-0.2-1.5-1.9-3.1-2.5-0.5-0.1-1.6-1.8-2.2-2.2-0.3-0.2-0.8-0.2-0.1-0.3-0.1-0.1-0.1-0.3 0-0.4-0.7-0.2-0.8-0.8-1.7-1.1-0.1-0.3 0.2-0.5 0-0.7-0.2-0.1-0.5 0.1-0.7 0-0.1 1.8 1.9 2 2.5 3.5-0.2-0.2-0.5-0.8-0.7-0.4-0.1 0.3 0.8 1.4 0.8 0.8-0.3 0.2-0.4-0.3-0.2-0.4 0.3 0.5 0.9 0.4 0.9 0.8-0.3 0.3-0.7-0.3-0.6 0.2 0.9 0.5 0.8 1 1.3 1.7-0.4-0.5-0.5 0.4-0.9 0-0.1-1.5-1.8-1.2-1.8-2.7-0.2-0.5-0.4-0.1-0.7-0.3 0.4-0.4-1.2-1.8-1.2-1.3-0.5-0.6-0.2-0.8-0.3-1.2-0.4-0.1-0.1 0.4-0.4 0.1 0.4-0.7-0.5-1.6-1.1-2.3l0.4-0.2c-0.7-0.4-0.8-1.2-1.3-2.1-0.6 0.9 1.2 2.9 0.9 3.3 0 0.1 0.7 0.9 0.9 1.3 0.8 1.4 1.4 3.1 2.5 3.9-0.2-0.1-0.1-0.7 0.2-0.4 0 0.3-0.2 0.6-0.2 0.9 0.6 0.6 1.1 1.1 1.7 1.7-1.3-0.8-2.6-1.6-3.5-2.8 0.5-0.1 1 0.8 1.4 0.7-1-0.8-0.7-1.1-1.1-2.1-0.4-0.3 0.1 0.4-0.3 0.5-1.3-1.2-2.3-1.9-2.9-3 0.1-0.1 0.2-0.2 0.2-0.4-0.1-0.1-0.1-0.1-0.2-0.2 0.6 0.6 0.8 1.1 1.5 1.2-0.3-0.3 0-0.4-0.4-0.7-0.2-0.4-0.5 0.2-0.7-0.1 0.2-0.3 0-0.7 0-1-0.2 0-0.4 0-0.5-0.3-0.2-0.1-0.3 0-0.4 0.2-0.1-0.1-0.1-0.3-0.2-0.4v-0.1c-0.1-0.1-0.2-0.3-0.3-0.4-0.8-2.1-1.3-4.9-2.4-6.7 0.3-1.2-0.2-3.6-0.5-4.2 0.3-0.5 0.1-2.8-0.1-3.2-0.5 0.2 0.1 1.3-0.4 1.6-0.5-2.2 0.3-4.7 0.1-6.7 0.2 0 0.4 0 0.4-0.1-0.3-0.3 0.2-0.9 0.1-1.6 0.4-0.9-0.9 1.3-0.9 2.2-0.1 0.2-0.5 0-0.4 0.5 0 0.7 0.2 2.1-0.2 3.1-0.2 0-0.2-0.2-0.4-0.1-0.1 0.5-0.2 1.1-0.3 1.6-0.1-0.2-0.2-0.5-0.2-0.6-0.5-2.5 0-5.4-0.2-7.5 0.4 0 0.4-0.6 0.7-0.8-0.1 0.8-0.3 1.8 0.2 2 0.4-3.9 1.2-8.2 3-12.4 1.3-3.1 3.2-6.2 5.5-8.7 0.2 0.3-0.1 1 0 1.3 2-2.1 3.3-4.6 5.5-6 0.1 0.5-0.9 0.9-0.9 1.3 0.9-0.9 1.2-0.6 2.2-0.9 0.3-0.4-0.4 0-0.5-0.4 0.6-0.4 1-0.9 1.4-1.2 0 0.1-0.1 0.1-0.1 0.2 1.1-0.3 1.2-0.9 2-1.1-0.1-0.1-0.3-0.2-0.6-0.1 0.2-0.1 0.5-0.3 0.8-0.4-0.1 0-0.1 0.1-0.2 0.1 0.6-0.3 0.1 0.1 0.1 0.3 0.1 0 0.2-0.1 0.2-0.1-0.6 0.4-1 0.6-1.2 1.3 0.3-0.2 0.4 0 0.8-0.3 0.4-0.2-0.1-0.5 0.2-0.7 0.3 0.2 0.7 0.1 1 0.1 0-0.2 0.1-0.4 0.3-0.5 0.1-0.2 0-0.3-0.1-0.3 0.1-0.1 0.2-0.1 0.4-0.2 0 0 0.1 0 0.1 0.1 1.7-0.8 3.6-1.5 5.4-2.1 1.1-0.1 2.5-0.1 3.1-0.8-0.2 0-0.4 0.1-0.5 0 0.1 0 0.3-0.1 0.4-0.1 0.2 0.2 0.5 0 0.2 0.4 0.1 0 0.1 0 0.2-0.1 0 0.1 0 0.1 0.1 0.2 0.1 0 0.2-0.1 0.3-0.3 0.1 0 0.2 0 0.3-0.1l-0.1 0.1c0.8-0.2 1.5-0.1 2.6-0.3 0-0.1 0-0.2 0-0.3 0.2 0 0.3-0.1 0.5-0.1v0.1c0.4-0.2 0.8 0.2 0.8 0 0.1 0 0.1 0 0.2 0 0 0.2-0.1 0.3-0.1 0.4 0.5 0 0.8 0 1-0.3-0.1-0.1-0.2-0.1-0.3-0.1 0.2 0 0.4-0.1 0.6-0.2 0.3 0.1 0.4 0.2 0.4 0.6 0.5-0.9 2.1 0.2 2.8 0 0.3 0 0.2-0.4 0.4-0.6 0.1 0 0.2 0 0.3 0 0.5 0.4 1.2 0.8 2.1 0.9-0.8-0.2 0.1-0.3 0.2-0.4-0.2 0-0.2-0.1-0.2-0.2 0.3 0 0.6 0.1 0.9 0.1 0 0.2 0.1 0.3 0.3 0.3 0.2 0 0.2-0.1 0.2-0.3 0.3 0 0.5 0 0.8 0.1-0.1 0.4 0.5 0.5 0.7 0.8-0.8-0.2-1.7-0.5-2 0 3.2 0.7 6.7 1.7 10.1 3.4 0 0 0 0 0 0.1h0.1c0.6 0.3 1.2 0.6 1.8 0.9 0.4 0.2 0.7 0.4 1.1 0.7 0 0.1 0 0.2 0.2 0.2h0.1c0.4 0.3 0.8 0.5 1.2 0.8 0.4 0.6 0.8 1.4 1.5 1.2 1.5 2.3 3.1 3.2 4.4 5.6 0.1-0.4 0.4 0.9 0.9 1-0.8-0.1 0.5 0.5-0.3 0.4 0.7 0.7 0.9 1 1 1.8-0.2-0.4-0.3-0.5-0.4-0.1-0.1 0.6 0.7-0.1 1 0.5-0.6 0-0.4 0.4-0.3 0.9 0.3-0.3 0.2-0.6 0.5-0.5 0.4 0.4-0.2 0.5-0.1 0.8 0.3-0.4 0.4-0.2 0.8 0-0.5 0.1 0.3 0.8-0.3 0.5 0.4 0.5 0.5 0.8 0.3 1.3-0.4-0.6-0.1-1-0.7-0.9 1.2 1 0.6 3.3 2 3.5-0.9 0.2 0 0.6-0.5 0.7 0.6 0.2 0.3 1 1 1.2-0.1 0.5-0.3 0.8-0.4 1.3 0.3 0.4 0.4 1.2 0.9 1.3-0.3 0.3-0.1 1.1-0.4 1.4 0.4-0.2 0.6-0.2 0.9 0.2-0.5 0.4-0.1 0.8-0.4 0.9 0.9 2.3-0.2 7 0.3 8.8-0.6-0.3-0.4 0.7-0.8 0.7 0.7 0.3 0.1 0.8 0.4 1.3-1.4 0.8 0 3-1.9 3.5 0.2 1.4-0.6 2-1.2 3.5-0.1-0.1-0.1-0.2-0.2-0.3-0.1 0.7-0.5 0.8-0.3 1.4l-0.4-0.2c0 1.1-1.7 1.7-2.1 3.4-0.1 0.5-1.6 1.8-2 2.5-0.2 0.3-0.1 0.8-0.3 0.1-0.1 0.1-0.3 0.1-0.4 0.1-0.1 0.7-0.7 0.9-0.9 1.8-0.2 0.1-0.5-0.2-0.7 0.1 0 0.2 0.2 0.5 0.1 0.7 1.8-0.1 1.8-2.2 3.2-2.9-0.1 0.3-0.7 0.6-0.3 0.8 0.3 0 1.2-1 0.7-0.9 0.3 0.2-0.3 0.4-0.4 0.3 0.5-0.4 0.3-0.9 0.7-1 0.3 0.2-0.3 0.7 0.3 0.6 0.4-0.9 0.9-1 1.5-1.5-0.4 0.4 0.5 0.5 0.1 0.9-1.5 0.3-1 2-2.5 2.2-0.4 0.3-0.1 0.4-0.2 0.7-0.5-0.3-1.6 1.4-1.2 1.3-0.5 0.5-0.8 0.3-1.2 0.4-0.1 0.4 0.4 0.1 0.2 0.4-0.7-0.3-1.6 0.7-2.2 1.4l-0.3-0.3c-0.3 0.8-1.1 0.9-2 1.5 1 0.5 2.7-1.6 3.2-1.3 0.1 0 0.9-0.8 1.2-1.1 1.3-1 2.9-1.8 3.5-2.9-0.1 0.2-0.7 0.2-0.4-0.1 0.3-0.1 0.6 0.1 0.9 0.1 1.6-2.2 3.6-4 5.1-6.6 0.1-0.1 0.2-0.3 0.3-0.5-0.2 0.6-0.5 1.1-0.7 1.7-1.6 1.1-2.2 4.3-4.1 4.8 0.1 1.1-0.9 1.4-1.4 2.7-0.3 0.4-0.5-0.3-0.8 0.2 0.3 1.4-2.5 1.3-2.4 2.6-0.2 0.1-0.5 0.2-0.7 0.4h-0.1c-0.1 0.1-0.1 0.1-0.2 0.2-0.5 0.3-0.9 0.6-1.3 0.9-1.4 0.8-3.3 1.4-4.8 2.2v-0.1c-0.4 0.2-0.9 0.1-0.8 0.5-0.3 0.2-0.6 0.4-0.9 0.6-0.1 0-0.1 0-0.2 0 0 0-0.1-0.1-0.2-0.1-0.2 0-0.2 0-0.3 0.1-1.3 0.1-3 0.7-3.5 1.1-0.5-0.3-2.8 0.2-3.2 0.4 0.2 0.3 0.8 0 1.3 0-0.3 0.1-0.6 0.2-0.7 0.4-0.1 0-0.1 0-0.2 0-0.1-0.3-0.4 0-0.5-0.1h0.1c-0.1-0.4-0.4-0.6-0.9-0.5 0 0.2 0 0.3-0.1 0.5 0.2 0 0.8-0.3 0.8 0-0.1 0-0.1 0.1 0 0.1-0.2 0.2-0.5 0.2-0.7 0.3-0.4 0-0.9 0-1.3 0-0.1-0.1-0.2-0.1-0.4-0.2-0.1 0.1-0.3 0.1-0.4 0.2-0.6 0-1.3 0.1-1.8 0.2 0-0.2 0-0.4-0.2-0.4-0.3 0.4-0.9-0.1-1.6 0.1-0.7-0.2 0.4 0.3 1.4 0.5-0.1 0.1-0.2 0.2-0.5 0.3M37.5 86.6c-1.4-1.3-4.3-1.9-4.8-2.9-0.8-0.1-2-1.6-3-1.8 0.5 0 0.1-0.3 0-0.5-0.2 0.2-0.4 0.3-0.6 0.3 0-0.8-0.8-1.3-1.5-1.8 0.5 0.3 0.9 0.5 1.5 0.6 0-0.2 0-0.4 0.2-0.4 0.7 0.7 1.5 1.4 2.4 2.1 1.2 1.1 7.4 3.9 8.5 4.4 0 0-0.2 0.4-0.1 0.4 0.5 0.3 0.9-0.3 1 0.4-1.6 0.2-2.2-0.9-3.6-0.8M83.6 63.4v0.1 0c-0.1 0 0-0.1 0-0.1M43.2 17.1c-0.1 0.1-0.2 0.1-0.2 0.1 0-0.1 0.1-0.1 0.2-0.1M41.3 17.6c0 0.1 0 0.1 0 0 0 0.1 0 0.1 0 0 0 0.1 0 0 0 0M34.7 19.7v0c0.1-0.1 0.1-0.1 0.2-0.1-0.1 0-0.2 0.1-0.2 0.1M50.7 16.4c-0.2 0-0.3 0-0.4-0.1 0.1 0 0.2 0 0.3 0h0.1c0 0 0 0 0 0.1M52.5 16.6c-0.1-0.1-0.4-0.2-0.7-0.2 0-0.1 0.1-0.2 0.1-0.2 0.1 0 0.2 0 0.3 0 0 0 0.1 0 0.1 0.1 0 0 0 0 0-0.1 0.1 0 0.1 0 0.2 0 0.1 0.2 0.2 0.3 0 0.4M55.8 16.9c-0.5 0.1-0.8-0.1-1.3-0.4 0.3 0 0.7 0.1 1 0.2 0.1 0 0.2 0.1 0.3 0.2M91.4 31.1c-0.2 0.1-0.3 0.1-0.5-0.3-1.6-3.4-4.5-8.3-8.7-12.4-4.1-4.2-9.6-7.6-14.6-9.4 0.3 0.1 0.7 0.3 0.5 0.5-4.4-1.7-8.6-2.7-12.6-3.1-10.8-1.7-22.1 0.6-30.8 6.2 0.1 0.2 0.1 0.3-0.2 0.5-3.2 2-7.7 5.5-11.3 10.1-3.6 4.6-6.4 10.4-7.6 15.6 0.1-0.3 0.2-0.8 0.4-0.6-2.6 10.3-1.9 19.1 1.7 26.7 0.2-0.2-0.5-1.2 0.1-1.1 0.4 1.1 0.9 2.3 1.5 3.4 0.7 1.6 1.5 3.7 2.6 4.2-2.8-5.5-4.7-11.8-4.9-17.4 0.4 0.8 0.3 2.5 0.9 3.2 0.2-3.8-0.7-8.9 0.6-12.5-0.3 1.9-0.5 4.5-0.3 7.1 0 1.9 0.9 7.2 1.5 7.7-0.1-0.1-0.1-0.6 0-0.6 0.3 0.6-0.4 0.9 0.1 1 0-0.2 0.1-0.2 0.2-0.2 1 2.2 2.3 5.6 3.3 7.3 1.2 2.2 2.8 4.2 4.6 6 0 0.1 0.1 0.1 0.1 0.2 0.2 0.5 0.9 1.1 1.1 1.5 0.1 0.2-0.1 0.5 0 0.7 1.5 2.8 3.7 5.6 6.4 7.8 1.7 2 4.4 4.1 7.7 5.9 3.2 1.7 6.9 3.1 10.3 3.9 2.4 0.6 5.3 0.7 7.1 1.1 2.4 0.5 6.6-0.2 8.7-0.6 12.6-1.9 25.1-10.5 31.4-22.9 6.5-12.5 6.3-27.8 0.7-39.5"
-                    fill-rule="nonzero"
-                    stroke="none"
-                    strokeWidth="1"
-                    strokeLinecap="butt"
-                    strokeLinejoin="miter"
-                    strokeMiterlimit="10"
-                    strokeDasharray=""
-                    strokeDashoffset="0"
-                    fontFamily="none"
-                    fontWeight="none"
-                    fontSize="none"
-                    textAnchor="none"
-                    data-fill-palette-color="tertiary"
-                  ></path>
-                </g>
-                <g transform="matrix(1,0,0,1,80.38303778829211,108)">
-                  <svg
-                    viewBox="0 0 159.23392442341577 144"
-                    height="144"
-                    width="159.23392442341577"
-                  >
-                    <g>
-                      <svg
-                        viewBox="0 0 159.23392442341577 144"
-                        height="144"
-                        width="159.23392442341577"
-                      >
-                        <g>
-                          <svg
-                            viewBox="0 0 160 144.69278505461432"
-                            height="144"
-                            width="159.23392442341577"
-                          >
-                            <g transform="matrix(1,0,0,1,0,58.54629984882165)">
-                              <svg
-                                viewBox="0 0 160 86.14648520579266"
-                                height="86.14648520579266"
-                                width="160"
-                              >
-                                <g id="textblocktransform">
-                                  <svg
-                                    viewBox="0 0 160 86.14648520579266"
-                                    height="86.14648520579266"
-                                    width="160"
-                                    id="textblock"
-                                  >
-                                    <g>
-                                      <svg
-                                        viewBox="0 0 160 71.47948611686698"
-                                        height="71.47948611686698"
-                                        width="160"
-                                      >
-                                        <g transform="matrix(1,0,0,1,0,0)">
-                                          <svg
-                                            width="160"
-                                            viewBox="122.30000000000001 -34.2 241.3 100.60000000000001"
-                                            height="71.47948611686698"
-                                            data-palette-color="#428dff"
-                                          >
-                                            <svg></svg>
-                                            <svg></svg>
-                                            <g
-                                              data-fill-palette-color="primary"
-                                              id="text-0"
-                                            >
-                                              <path
-                                                className="fill-primary-light dark:fill-primary-dark"
-                                                transform="translate(163.45,0)"
-                                                d="M27.1-32.2v-1.2h10.1v1.2l-3.5 1.2v31h-3.4l-22.25-27.4v24.95l4.2 1.25v1.2h-10v-1.2l3.65-1.25v-29.3l-4.35-0.45v-1.2h9.4l20 24.9v-22.55zM56.95-17.45v0c0-4.067-1.417-6.1-4.25-6.1v0c-2.967 0-4.767 2.6-5.4 7.8v0h9.55c0.067-0.667 0.1-1.233 0.1-1.7zM52.25 0.55v0c-3.567 0-6.46-1.167-8.68-3.5-2.213-2.333-3.32-5.417-3.32-9.25 0-3.833 1.157-6.917 3.47-9.25 2.32-2.333 5.263-3.5 8.83-3.5 3.567 0 6.233 0.757 8 2.27 1.767 1.52 2.65 3.63 2.65 6.33v0c0 0.767-0.067 1.567-0.2 2.4v0h-15.85v1.75c0 3.733 0.467 6.557 1.4 8.47 0.933 1.92 2.34 2.88 4.22 2.88 1.887 0 3.603-0.39 5.15-1.17 1.553-0.787 2.763-1.897 3.63-3.33v0c0.733 0.233 1.1 0.583 1.1 1.05v0c0 0.833-0.783 1.8-2.35 2.9v0c-0.767 0.533-1.86 0.99-3.28 1.37-1.413 0.387-3.003 0.58-4.77 0.58zM75.2-1.25v1.25h-9.6v-1.25l3.5-0.95 7.7-9.1-9.1-11.6-2.4-0.25v-1.25h12.1l-0.05 1.2-2.9 0.4 6.5 8.45 6.5-7.85-3.85-0.95v-1.25h9.6v1.25l-3.5 0.95-7.75 9.3 8.75 11.4 2.85 0.25v1.25h-12.55v-1.15l2.9-0.4-6.15-8.2-6.25 7.55zM104.7 0.55v0c-3.2 0-5.233-1.117-6.1-3.35v0c-0.367-0.967-0.55-2.183-0.55-3.65v0c0-5.533 0.083-11.083 0.25-16.65v0h-3.4v-1.3h3.45l0.05-7.35h6.4l-0.15 7.35h7v1.3h-7l-0.1 4.05c-0.1 5.4-0.15 10.35-0.15 14.85v0c0 1.3 0.143 2.183 0.43 2.65 0.28 0.467 0.797 0.7 1.55 0.7 0.747 0 1.58-0.527 2.5-1.58 0.913-1.047 1.57-2.303 1.97-3.77v0c0.667 0 1 0.233 1 0.7v0c0 0.1-0.017 0.217-0.05 0.35v0c-0.533 1.733-1.44 3.117-2.72 4.15-1.287 1.033-2.747 1.55-4.38 1.55zM117.9-2.95v0-28.8l-4.35-0.5v-1.15h15.35v1.15l-4.35 0.5v29.3c0 4.9-1.057 8.593-3.17 11.08-2.12 2.48-5.03 3.72-8.73 3.72v0c-2.667 0-4.79-0.423-6.37-1.27-1.587-0.853-2.38-2.087-2.38-3.7 0-1.62 0.583-2.763 1.75-3.43v0c0.633-0.333 1.277-0.5 1.93-0.5 0.647 0 1.303 0.15 1.97 0.45v0c-0.167 1.133-0.25 2.167-0.25 3.1 0 0.933 0.277 1.81 0.83 2.63 0.547 0.813 1.27 1.22 2.17 1.22 0.9 0 1.643-0.167 2.23-0.5 0.58-0.333 1.053-0.74 1.42-1.22 0.367-0.487 0.677-1.147 0.93-1.98 0.247-0.833 0.437-1.59 0.57-2.27 0.133-0.687 0.233-1.58 0.3-2.68v0c0.1-1.667 0.15-3.383 0.15-5.15zM134-1.7v0c-1.4-0.9-2.1-2.05-2.1-3.45v0c0-1.933 1.533-3.133 4.6-3.6v0c0.533 3.1 1.667 5.25 3.4 6.45v0c1.533 1.067 3.157 1.6 4.87 1.6 1.72 0 3.203-0.523 4.45-1.57 1.253-1.053 1.88-2.78 1.88-5.18v0c0-1.8-0.643-3.227-1.93-4.28-1.28-1.047-2.903-1.97-4.87-2.77-1.967-0.8-3.3-1.36-4-1.68-0.7-0.313-1.59-0.787-2.67-1.42-1.087-0.633-1.897-1.233-2.43-1.8v0c-1.333-1.533-2-3.3-2-5.3v0c0-3.133 1.167-5.5 3.5-7.1 2.333-1.6 5.583-2.4 9.75-2.4v0c2.367 0 5.117 0.267 8.25 0.8v0 7.4h-1.65l-2-5.9c-1.667-0.533-3.36-0.8-5.08-0.8-1.713 0-3.203 0.54-4.47 1.62-1.267 1.087-1.9 2.697-1.9 4.83v0c0 2.433 1.1 4.25 3.3 5.45v0c1.1 0.567 2.4 1.133 3.9 1.7 1.5 0.567 2.583 0.993 3.25 1.28 0.667 0.28 1.55 0.727 2.65 1.34 1.1 0.62 1.917 1.23 2.45 1.83 0.533 0.6 1.017 1.39 1.45 2.37 0.433 0.987 0.65 2.063 0.65 3.23v0c0 2.833-1.2 5.15-3.6 6.95-2.4 1.8-5.677 2.7-9.83 2.7-4.147 0-7.42-0.767-9.82-2.3z"
-                                                fill-rule="nonzero"
-                                                stroke="none"
-                                                strokeWidth="1"
-                                                strokeLinecap="butt"
-                                                strokeLinejoin="miter"
-                                                strokeMiterlimit="10"
-                                                strokeDasharray=""
-                                                strokeDashoffset="0"
-                                                fontFamily="none"
-                                                fontWeight="none"
-                                                fontSize="none"
-                                                textAnchor="none"
-                                                data-fill-palette-color="primary"
-                                                opacity="1"
-                                              ></path>
-                                              <path
-                                                className="fill-primary-light dark:fill-primary-dark"
-                                                transform="translate(120.65,53.9)"
-                                                d="M38.1-16.75v0c0 11.167-6.933 16.75-20.8 16.75v0h-15.65v-1.2l4.3-0.5v-29.95l-4.3-0.6v-1.3c1.533 0.1 3.317 0.15 5.35 0.15v0h4.8l6.2-0.2c3.033 0 5.693 0.25 7.98 0.75 2.28 0.5 4.37 1.35 6.27 2.55v0c3.9 2.467 5.85 6.983 5.85 13.55zM30.55-16.7v0c0-5.767-1.083-9.767-3.25-12v0c-0.7-0.767-1.35-1.343-1.95-1.73-0.6-0.38-1.35-0.67-2.25-0.87v0c-1.233-0.3-2.683-0.483-4.35-0.55v0h-0.75c-2.433 0-4.217 0.083-5.35 0.25v0 29.85l4.65 0.15c3.167-0.133 5.517-0.567 7.05-1.3v0c4.133-1.967 6.2-6.567 6.2-13.8zM59.05-17.45v0c0-4.067-1.417-6.1-4.25-6.1v0c-2.967 0-4.767 2.6-5.4 7.8v0h9.55c0.067-0.667 0.1-1.233 0.1-1.7zM54.35 0.55v0c-3.567 0-6.457-1.167-8.67-3.5-2.22-2.333-3.33-5.417-3.33-9.25 0-3.833 1.16-6.917 3.48-9.25 2.313-2.333 5.253-3.5 8.82-3.5 3.567 0 6.233 0.757 8 2.27 1.767 1.52 2.65 3.63 2.65 6.33v0c0 0.767-0.067 1.567-0.2 2.4v0h-15.85v1.75c0 3.733 0.467 6.557 1.4 8.47 0.933 1.92 2.343 2.88 4.23 2.88 1.88 0 3.597-0.39 5.15-1.17 1.547-0.787 2.753-1.897 3.62-3.33v0c0.733 0.233 1.1 0.583 1.1 1.05v0c0 0.833-0.783 1.8-2.35 2.9v0c-0.767 0.533-1.857 0.99-3.27 1.37-1.42 0.387-3.013 0.58-4.78 0.58zM83.15-23.15v-1.25h9.6v1.2l-3.55 1.05-8.1 22.15h-2.7l-9.35-22.85-2.25-0.35v-1.2h11.3v1.2l-2.95 0.35 6.4 16.25 5.55-15.55zM110.45-17.45v0c0-4.067-1.417-6.1-4.25-6.1v0c-2.967 0-4.767 2.6-5.4 7.8v0h9.55c0.067-0.667 0.1-1.233 0.1-1.7zM105.75 0.55v0c-3.567 0-6.457-1.167-8.67-3.5-2.22-2.333-3.33-5.417-3.33-9.25 0-3.833 1.16-6.917 3.48-9.25 2.313-2.333 5.253-3.5 8.82-3.5 3.567 0 6.233 0.757 8 2.27 1.767 1.52 2.65 3.63 2.65 6.33v0c0 0.767-0.067 1.567-0.2 2.4v0h-15.85v1.75c0 3.733 0.467 6.557 1.4 8.47 0.933 1.92 2.343 2.88 4.23 2.88 1.88 0 3.597-0.39 5.15-1.17 1.547-0.787 2.753-1.897 3.62-3.33v0c0.733 0.233 1.1 0.583 1.1 1.05v0c0 0.833-0.783 1.8-2.35 2.9v0c-0.767 0.533-1.857 0.99-3.27 1.37-1.42 0.387-3.013 0.58-4.78 0.58zM132.5-1.15v1.15h-12.9v-1.15l3.25-0.3v-34.45l-4.3-0.15v-1.2l10.7-1.15v36.95zM158.05-3.05c-2.333 2.4-5.517 3.6-9.55 3.6-4.033 0-7.217-1.2-9.55-3.6-2.333-2.4-3.5-5.45-3.5-9.15 0-3.7 1.167-6.75 3.5-9.15 2.333-2.4 5.517-3.6 9.55-3.6 4.033 0 7.217 1.2 9.55 3.6 2.333 2.4 3.5 5.45 3.5 9.15 0 3.7-1.167 6.75-3.5 9.15zM142.4-12.2c0 7.7 2.033 11.55 6.1 11.55 4.067 0 6.1-3.85 6.1-11.55v0c0-3.767-0.45-6.627-1.35-8.58-0.9-1.947-2.483-2.92-4.75-2.92-2.267 0-3.85 0.973-4.75 2.92-0.9 1.953-1.35 4.813-1.35 8.58zM182.45-24.95v0c2.933 0 5.39 0.957 7.37 2.87 1.987 1.92 2.98 4.953 2.98 9.1 0 4.153-1.16 7.453-3.48 9.9-2.313 2.453-5.637 3.68-9.97 3.68v0c-1.933 0-3.583-0.45-4.95-1.35v0 11.8l3.45 0.3v1.15h-12.9v-1.15l3.05-0.3v-33.7l-4.35-0.25v-1.15l10.6-0.65-0.15 4.2c1.833-2.967 4.617-4.45 8.35-4.45zM174.4-17.25v0 14.7c1.333 1.1 2.85 1.65 4.55 1.65 1.7 0 3-0.373 3.9-1.12 0.9-0.753 1.55-1.813 1.95-3.18v0c0.7-2.233 1.05-5.133 1.05-8.7v0c0-6.333-1.633-9.5-4.9-9.5v0c-1.733 0-3.257 0.7-4.57 2.1-1.32 1.4-1.98 2.75-1.98 4.05zM213.45-17.45v0c0-4.067-1.417-6.1-4.25-6.1v0c-2.967 0-4.767 2.6-5.4 7.8v0h9.55c0.067-0.667 0.1-1.233 0.1-1.7zM208.75 0.55v0c-3.567 0-6.457-1.167-8.67-3.5-2.22-2.333-3.33-5.417-3.33-9.25 0-3.833 1.16-6.917 3.48-9.25 2.313-2.333 5.253-3.5 8.82-3.5 3.567 0 6.233 0.757 8 2.27 1.767 1.52 2.65 3.63 2.65 6.33v0c0 0.767-0.067 1.567-0.2 2.4v0h-15.85v1.75c0 3.733 0.467 6.557 1.4 8.47 0.933 1.92 2.343 2.88 4.23 2.88 1.88 0 3.597-0.39 5.15-1.17 1.547-0.787 2.753-1.897 3.62-3.33v0c0.733 0.233 1.1 0.583 1.1 1.05v0c0 0.833-0.783 1.8-2.35 2.9v0c-0.767 0.533-1.857 0.99-3.27 1.37-1.42 0.387-3.013 0.58-4.78 0.58zM232.5-18.5c0.567-1.8 1.61-3.333 3.13-4.6 1.513-1.267 2.897-1.9 4.15-1.9 1.247 0 2.087 0.307 2.52 0.92 0.433 0.62 0.65 1.13 0.65 1.53v0c0 1.367-0.5 2.35-1.5 2.95v0c-0.667-0.533-1.657-0.8-2.97-0.8-1.32 0-2.63 0.407-3.93 1.22-1.3 0.82-1.95 1.713-1.95 2.68v0 15.05l4.5 0.3v1.15h-13.95v-1.15l3.3-0.25v-21.35l-4.05-0.2v-1.15l10.2-0.55-0.3 6.15z"
-                                                fillRule="nonzero"
-                                                stroke="none"
-                                                strokeWidth="1"
-                                                strokeLinecap="butt"
-                                                strokeLinejoin="miter"
-                                                strokeMiterlimit="10"
-                                                strokeDasharray=""
-                                                strokeDashoffset="0"
-                                                fontFamily="none"
-                                                fontWeight="none"
-                                                fontSize="none"
-                                                textAnchor="none"
-                                                data-fill-palette-color="primary"
-                                                opacity="1"
-                                              ></path>
-                                            </g>
-                                          </svg>
-                                        </g>
-                                      </svg>
-                                    </g>
-                                    <g transform="matrix(1,0,0,1,0,76.40449670949027)">
-                                      <svg
-                                        viewBox="0 0 160 9.741988496302385"
-                                        height="9.741988496302385"
-                                        width="160"
-                                      >
-                                        <g transform="matrix(1,0,0,1,0,0)">
-                                          <svg
-                                            className="fill-primary-light dark:fill-primary-dark"
-                                            width="160"
-                                            viewBox="-1.3 -36.35 608.4499999999999 37.050000000000004"
-                                            height="9.741988496302385"
-                                            data-palette-color="#428dff"
-                                          >
-                                            <path
-                                              className="fill-primary-light dark:fill-primary-dark"
-                                              d="M28.25 0L18.75 0 19.05-7.85 10.7-7.85 8.25 0-1.3 0 11.55-36 28.05-36 28.25 0ZM19.85-27.25L16.7-27.25 12.75-14.55 19.3-14.55 19.85-27.25ZM57.2-7.8L55.8 0 33 0 39.35-36 48.35-36 43.4-7.8 57.2-7.8ZM86.3-7.8L84.9 0 59.8 0 66.15-36 91.25-36 89.85-28.2 73.75-28.2 72.6-21.45 85.9-21.45 84.5-13.65 71.2-13.65 70.2-7.8 86.3-7.8ZM117.24-36L127.49-36 113.39-17.85 121.14 0 110.89 0 105.99-11.5 97.74 0 87.49 0 100.74-17.85 93.84-36 104.09-36 108.24-24.3 117.24-36ZM152.34 0L142.84 0 143.14-7.85 134.79-7.85 132.34 0 122.79 0 135.64-36 152.14-36 152.34 0ZM143.94-27.25L140.79-27.25 136.84-14.55 143.39-14.55 143.94-27.25ZM182.19-36L191.19-36 184.89 0 175.84 0 169.54-19.7 166.14 0 157.09 0 163.44-36 172.74-36 178.94-17.6 182.19-36ZM198.29-36L212.29-36Q218.84-36 221.64-32.7L221.64-32.7Q223.69-30.2 223.69-26.45L223.69-26.45Q223.69-25.1 223.44-23.6L223.44-23.6 221.49-12.45Q220.49-6.8 216.59-3.4 212.69 0 206.49 0L206.49 0 191.94 0 198.29-36ZM209.14-28.2L209.14-28.2 205.89-28.2 202.34-7.8 207.09-7.8Q209.44-7.8 210.76-9.08 212.09-10.35 212.59-13.25L212.59-13.25 214.04-21.45Q214.24-22.55 214.24-23.45L214.24-23.45Q214.24-28.2 209.14-28.2ZM251.88-7.8L250.48 0 225.38 0 231.73-36 256.83-36 255.43-28.2 239.33-28.2 238.18-21.45 251.48-21.45 250.08-13.65 236.78-13.65 235.78-7.8 251.88-7.8ZM262.58-36L277.08-36Q282.78-36 285.18-33.5L285.18-33.5Q286.98-31.6 286.98-28.7L286.98-28.7Q286.98-27.7 286.78-26.6L286.78-26.6 286.08-22.7Q285.48-19.45 284.01-17.45 282.53-15.45 280.73-14.6L280.73-14.6 285.28 0 275.18 0 270.98-13.45 267.88-13.45 265.83 0 256.23 0 262.58-36ZM273.93-28.2L273.93-28.2 270.18-28.2 268.88-19.5 272.38-19.5Q275.03-19.5 276.18-20.78 277.33-22.05 277.33-24.15L277.33-24.15Q277.33-28.2 273.93-28.2ZM307.38-36L316.78-36 322.13-18.75 333.63-36 343.08-36 336.73 0 327.68 0 331.28-20.3 323.58-8.2 317.03-8.2 313.63-20.3 310.03 0 301.03 0 307.38-36ZM356.47 0.7L356.47 0.7Q349.32 0.7 346.57-2.4L346.57-2.4Q344.67-4.6 344.67-8.55L344.67-8.55Q344.67-10.25 345.02-12.2L345.02-12.2 347.07-23.7Q349.32-36.35 363.32-36.35L363.32-36.35Q369.12-36.35 372.12-33.88 375.12-31.4 375.12-27.1L375.12-27.1Q375.12-25.1 374.55-23.13 373.97-21.15 373.42-20.2L373.42-20.2 372.87-19.2 364.02-19.2Q365.67-21.1 365.67-24.6L365.67-24.6Q365.67-26.45 364.72-27.53 363.77-28.6 361.92-28.6L361.92-28.6Q356.87-28.6 355.92-23L355.92-23 354.17-12.9Q353.92-11.6 353.92-10.35L353.92-10.35Q353.92-7.05 357.87-7.05L357.87-7.05Q362.62-7.05 363.57-12.4L363.57-12.4 372.57-12.4 372.32-10.95Q370.32 0.7 356.47 0.7ZM412.37-7.8L410.97 0 388.17 0 394.52-36 403.52-36 398.57-7.8 412.37-7.8ZM442.27 0L432.77 0 433.07-7.85 424.72-7.85 422.27 0 412.72 0 425.57-36 442.07-36 442.27 0ZM433.87-27.25L430.72-27.25 426.77-14.55 433.32-14.55 433.87-27.25ZM459.66 0.7L459.66 0.7Q452.51 0.7 449.76-2.4L449.76-2.4Q447.86-4.6 447.86-8.55L447.86-8.55Q447.86-10.25 448.21-12.2L448.21-12.2 450.26-23.7Q452.51-36.35 466.51-36.35L466.51-36.35Q472.31-36.35 475.31-33.88 478.31-31.4 478.31-27.1L478.31-27.1Q478.31-25.1 477.74-23.13 477.16-21.15 476.61-20.2L476.61-20.2 476.06-19.2 467.21-19.2Q468.86-21.1 468.86-24.6L468.86-24.6Q468.86-26.45 467.91-27.53 466.96-28.6 465.11-28.6L465.11-28.6Q460.06-28.6 459.11-23L459.11-23 457.36-12.9Q457.11-11.6 457.11-10.35L457.11-10.35Q457.11-7.05 461.06-7.05L461.06-7.05Q465.81-7.05 466.76-12.4L466.76-12.4 475.76-12.4 475.51-10.95Q473.51 0.7 459.66 0.7ZM504.46-36L513.46-36 507.16 0 498.11 0 500.41-13 490.66-13 488.41 0 479.36 0 485.71-36 494.71-36 492.01-20.8 501.76-20.8 504.46-36ZM538.41-7.8L537.01 0 514.21 0 520.56-36 529.56-36 524.61-7.8 538.41-7.8ZM568.3 0L558.8 0 559.1-7.85 550.75-7.85 548.3 0 538.75 0 551.6-36 568.1-36 568.3 0ZM559.9-27.25L556.75-27.25 552.8-14.55 559.35-14.55 559.9-27.25ZM598.15-36L607.15-36 600.85 0 591.8 0 585.5-19.7 582.1 0 573.05 0 579.4-36 588.7-36 594.9-17.6 598.15-36Z"
-                                              opacity="1"
-                                              transform="matrix(1,0,0,1,0,0)"
-                                              data-fill-palette-color="secondary"
-                                              id="text-1"
-                                            ></path>
-                                          </svg>
-                                        </g>
-                                      </svg>
-                                    </g>
-                                  </svg>
-                                </g>
-                              </svg>
-                            </g>
-                            <g transform="matrix(1,0,0,1,30.560000000000002,0)">
-                              <svg
-                                viewBox="0 0 98.88 53.62128925619835"
-                                height="53.62128925619835"
-                                width="98.88"
-                              >
-                                <g>
-                                  <svg
-                                    xmlns="http://www.w3.org/2000/svg"
-                                    data-name="Layer 1"
-                                    viewBox="13.2 29.5 72.6 39.370000000000005"
-                                    x="0"
-                                    y="0"
-                                    height="53.62128925619835"
-                                    width="98.88"
-                                    data-fill-palette-color="accent"
-                                    id="icon-0"
-                                  >
-                                    <polygon
-                                      className="fill-primary-light dark:fill-primary-dark"
-                                      points="39.6 68.87 42.82 68.87 59.4 29.5 56.18 29.5 39.6 68.87"
-                                      data-fill-palette-color="accent"
-                                    ></polygon>
-                                    <polygon
-                                      className="fill-primary-light dark:fill-primary-dark"
-                                      points="66 29.5 66 34.14 81.13 49.17 66 64.21 66 68.84 85.8 49.17 66 29.5"
-                                      data-fill-palette-color="accent"
-                                    ></polygon>
-                                    <polygon
-                                      className="fill-primary-light dark:fill-primary-dark"
-                                      points="33 68.84 33 64.21 17.87 49.17 33 34.14 33 29.5 13.2 49.17 33 68.84"
-                                      data-fill-palette-color="accent"
-                                    ></polygon>
-                                  </svg>
-                                </g>
-                              </svg>
-                            </g>
-                            <g></g>
-                          </svg>
-                        </g>
-                      </svg>
-                    </g>
-                  </svg>
-                </g>
-              </svg>
-            </g>
-            <defs></defs>
-          </svg>
-          <rect
-            width="320"
-            height="360"
-            fill="none"
-            stroke="none"
-            visibility="hidden"
-          ></rect>
-        </g>
-      </svg>
-    </g>
-  </svg>
-);
